moved hypothesis tests inside package
